+++ flash/core/integrations/labelstudio/data_source.py
@@ -8,16 +8,14 @@ from pytorch_lightning.utilities.cloud_io import get_filesystem
 
 from flash.core.data.auto_dataset import AutoDataset, IterableAutoDataset
 from flash.core.data.data_source import DataSource, DefaultDataKeys, has_len
-from flash.core.utilities.imports import _PYTORCHVIDEO_AVAILABLE, _TEXT_AVAILABLE, _TORCHVISION_AVAILABLE
+from flash.core.utilities.imports import _PYTORCHVIDEO_AVAILABLE, _TORCHVISION_AVAILABLE
 from flash.core.utilities.stages import RunningStage
 
 if _TORCHVISION_AVAILABLE:
     from torchvision.datasets.folder import default_loader
 
-if _TEXT_AVAILABLE:
-    from flash.text.data import TokenizerState
-    from flash.text.tokenizers.base import BaseTokenizer
-
+# from flash.text.data import TokenizerState
+# from flash.text.tokenizers import BaseTokenizer
 
 DATA_TYPE = TypeVar("DATA_TYPE")
 
@@ -194,30 +192,6 @@ class LabelStudioImageClassificationDataSource(LabelStudioDataSource):
         # loading image
         image = default_loader(p)
         result = {DefaultDataKeys.INPUT: image, DefaultDataKeys.TARGET: self._get_labels_from_sample(sample["label"])}
-        return result
-
-
-class LabelStudioTextClassificationDataSource(LabelStudioDataSource):
-    """The ``LabelStudioTextDataSource`` expects the input to
-    :meth:`~flash.core.data.data_source.DataSource.load_data` to be a json export from label studio.
-    Export data should point to text data
-    """
-
-    def __init__(self, tokenizer: BaseTokenizer):
-        super().__init__()
-        self.set_state(TokenizerState(tokenizer))
-
-    def load_sample(self, sample: Mapping[str, Any] = None, dataset: Optional[Any] = None) -> Any:
-        """Load 1 sample from dataset."""
-        data = ""
-        for key in sample.get("data"):
-            data += sample.get("data").get(key)
-        tokenized_data = self.get_state(TokenizerState).tokenizer(data)
-        for key in tokenized_data:
-            tokenized_data[key] = torch.tensor(tokenized_data[key])
-        tokenized_data["labels"] = self._get_labels_from_sample(sample["label"])
-        # separate text data type block
-        result = tokenized_data
         return result
