+++ flash_examples/image_classification.py
@@ -1,3 +1,5 @@
+import pytorch_lightning as pl
+pl.seed_everything(42)
 # Copyright The PyTorch Lightning team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
